requirements.txt
google-generativeai>=0.8.0
gradio>=4.0.0
python-docx>=1.2.0
Pillow
numpy